Journey Mode 1.8.1: fix deposit-screen freeze, stale-jar sound, and shift-click stack sizes
1.8.0 shipped a game-breaking defect: on modpacks whose crafting graph contains
a cycle, the Deposit tab re-resolved an item threshold on every rendered frame,
pinning the render thread at 100% with no crash report and nothing in the log.
A single reverse-crafting recipe was enough to trigger it, since ordinary vanilla
doors/boats/signs close the loop from the other side.

Three compounding causes, all fixed:
- RecipeDepthCalculator now memoizes cycle-tainted depths. This reverses 1.8.0s
  "never cache a tainted result" rule, which protected a theoretical property
  (purely structural depths) at the cost of a real one (bounded work). Tainted
  items are tracked in provisionalDepths so the tradeoff stays visible.
- The GUI and tooltip layers cache the resolved threshold per item and per
  ConfigHandler rules generation, instead of asking once per frame.
- The walk itself is bounded: MAX_RECURSION_DEPTH closes a genuine
  StackOverflowError path, MAX_NODE_VISITS caps total work, and two exact early
  exits cut the common case short.

Also fixed:
- Four of the nine 1.8.0 jars (Forge 1.12.2/1.16.5/1.20.1, NeoForge 1.21.1) were
  staged before the unlock sound was removed and shipped it. Source was already
  clean; all nine are rebuilt here. JM_Project.md gains a rebuild-discipline note.
- Shift-click withdrawal used a flat 64 on both ends, so a size-1 item came back
  as 64 slots worth. Both the screen and the server-authoritative packet handler
  now clamp to the items own stack limit.
- Team badge and screen title were drawn on the same line; Stats per-mod list was
  capped at 6 rows with no clearance check and overran the "Latest:" footer and
  the vanilla inventory label. Both now derive their positions from one constant,
  and the Stats list scrolls.
- Forge 1.19.2 never called the 6-arg updateFromSync, so it had no unlock message
  and no team badge at all.
- Deposit All emitted one chat line per unlocked item on top of the batched
  action-bar message; now a single summary line.

Audit fixes (no reported symptom, found while verifying the above):
- Journey search, Stats and tooltips no longer rebuild their whole view per frame.
- The sync packet sheds optional payload before it can exceed vanillas 1 MiB
  custom-payload ceiling, which would have kicked the client on every join.
- Progress percentage uses long math at all 27 sites.

All 9 environments compile and build independently. Full account in
1.8.0_POST_RELEASE_FINDINGS.md.

+++ Forge_1_12_2/src/main/java/com/aryangpt007/journeymode/data/TeamData.java
@@ -88,7 +88,9 @@ public class TeamData {
         int count = getCollectedCount(stack);
         int threshold = getThreshold(stack.getItem());
         if (threshold <= 0) return 100;
-        return Math.min(100, (count * 100) / threshold);
+        // long math: a collected count past ~21M would overflow int and report a negative
+        // percentage. Thresholds are always >= 1 (calculateThreshold clamps), so no /0 here.
+        return (int) Math.min(100L, (long) count * 100L / threshold);
     }
 
     public Set<String> getUnlockedItems() {
